Warning fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@31339 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shaderplugins/glshader_cg/glshader_cgcommon_clip.cpp
@@ -74,7 +74,7 @@ void csShaderGLCGCommon::ClipsToVmap ()
     }
     else
     {
-      CS::ShaderVarStringID svName (svClipPlane - c);
+      CS::ShaderVarStringID svName (svClipPlane - (uint)c);
       clipPlane[c].AttachNew (new csShaderVariable (svName));
       clipPlane[c]->SetValue (0.0f);
       vme.name = svName;
